+++ test/resolver/x_router_resolver_test.dart
@@ -28,11 +28,18 @@ void main() {
 
     group('built ins =>', () {
       test('RedirectResolver', () async {
-        final redirectResolver = XRedirectResolver(from: '/', to: '/dashboard');
+        bool shouldRedirect = true;
+        final redirectResolver = XRedirectResolver(
+          from: '/',
+          to: '/dashboard',
+          when: () => shouldRedirect,
+        );
         final redirectWithParamsResolver =
             XRedirectResolver(from: '/products/:id', to: '/products/:id/info');
         expect(redirectResolver.resolve('/'),
             equals(const Redirect('/dashboard')));
+        shouldRedirect = false;
+        expect(redirectResolver.resolve('/'), equals(const Next()));
         expect(redirectResolver.resolve('/other'), equals(const Next()));
         expect(redirectWithParamsResolver.resolve('/products/123'),
             equals(const Redirect('/products/123/info')));